% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skimmers.R
\name{get_skimmers.sfc_MULTIPOLYGON}
\alias{get_skimmers.sfc_MULTIPOLYGON}
\title{Retrieve the summary functions for a specific data type}
\usage{
get_skimmers.sfc_MULTIPOLYGON(column)
}
\description{
Retrieve the summary functions for a specific data type
}
