% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skim_to_var_table}
\alias{skim_to_var_table}
\title{Use skimr to produce a variable table}
\usage{
skim_to_var_table(x, digits = 2)
}
\arguments{
\item{digits}{Number of digits to use for meta data}
}
\description{
Use skimr to produce a variable table
}
