% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_with.R
\name{skim_var_table}
\alias{skim_var_table}
\title{Alternative skim_with() functions built with the skimr function factory.
Skimmers to produce a table of variables with same named values
in a single column.}
\usage{
skim_var_table(data, ...)
}
\description{
Alternative skim_with() functions built with the skimr function factory.
Skimmers to produce a table of variables with same named values
in a single column.
}
